ProyectoExamen/examen/requirements.txt
django == 4.0.6
pillow 
django-admin-interface
django-crispy-forms
django-rest-framework